MPI_COUNT & AINT: Tie MPI_Count to MPC_Count and use size_t

+++ mpcframework/MPC_Message_Passing/include/mpcmp.h
@@ -63,8 +63,8 @@ extern "C"
 
   typedef unsigned int mpc_pack_indexes_t;
   typedef long mpc_pack_absolute_indexes_t;
-  typedef long MPC_Aint;
-  typedef long MPC_Count;
+  typedef size_t MPC_Aint;
+  typedef size_t MPC_Count;
   typedef sctk_communicator_t MPC_Comm;
 
   typedef struct {
